backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.17
pytesseract==0.3.13
opencv-python==4.10.0.84
PyMuPDF==1.24.14
Pillow==11.0.0
python-magic==0.4.27
pdf2image==1.17.0
pydantic==2.10.3
python-dotenv==1.0.1
supabase==2.8.0
asyncpg==0.29.0
sqlalchemy==2.0.23